+++ source/Energinet.DataHub.MeteringPoints.Domain/BusinessProcesses/UpdateMasterData/UpdateMasterDataPolicies.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MeteringPoints.Domain.BusinessProcesses
+namespace Energinet.DataHub.MeteringPoints.Domain.BusinessProcesses.UpdateMasterData
 {
     public record UpdateMasterDataPolicies(int NumberOfDaysEffectiveDateIsAllowedToBeforeToday = 0);
 }
